Enhance feature link (#3075)
* Fix feature_links edge cases

* Add basic FeatureLinksAPI

* Add feature-link to enhance autolink and normal url in feature detail page

* Add tooltip to enhanceChromeStatusLink, polish styles

* Make featureLinks non blocking request

* Add closedTimestamp in enhanceChromeStatusLink

* Remove placeholder function

* Add back white-space: pre-wrap and fix CSS

* Use gray color badge for closed issue

* Rename fallback and change it to a lit-html object

* Move feature link summary into tooltip

* Support custom text in enhanceUrl

+++ internals/feature_links.py
@@ -60,7 +60,7 @@ def _index_link(link: Link, fe: FeatureEntry) -> None:
   feature_id = fe.key.integer_id()
   feature_links = FeatureLinks.query(FeatureLinks.url == link.url).fetch(None)
   feature_link: FeatureLinks = feature_links[0] if feature_links else None
-  if feature_link:
+  if hasattr(feature_link, 'feature_ids'):
     if feature_id not in feature_link.feature_ids:
       feature_link.feature_ids.append(feature_id)
       feature_link.type = link.type
@@ -87,20 +87,32 @@ def _remove_link(link: Link, fe: FeatureEntry) -> None:
   feature_id = fe.key.integer_id()
   feature_links = FeatureLinks.query(FeatureLinks.url == link.url).fetch(None)
   feature_link: FeatureLinks = feature_links[0] if feature_links else None
-  if feature_id in feature_link.feature_ids:
-     feature_link.feature_ids.remove(feature_id)
-     if feature_link.feature_ids:
-       feature_link.put()
-       logging.info(f'Updated indexed link {link.url}')
-     else:
-       # delete the link if it is not used by any feature
-       feature_link.key.delete()
-       logging.info(f'Delete indexed link {link.url}')
+  if hasattr(feature_link, 'feature_ids'):
+    if feature_id in feature_link.feature_ids:
+      feature_link.feature_ids.remove(feature_id)
+      if feature_link.feature_ids:
+        feature_link.put()
+        logging.info(f'Updated indexed link {link.url}')
+      else:
+        # delete the link if it is not used by any feature
+        feature_link.key.delete()
+        logging.info(f'Delete indexed link {link.url}')
 
 
-def _get_feature_links(fe: FeatureEntry) -> list[Link]:
-  """Return a list of Link in the given feature entry."""
-  pass
+def _get_feature_links(feature_id: int) -> list[FeatureLinks]:
+  """Return a list of FeatureLinks for a given feature id"""
+  feature_links = FeatureLinks.query(
+      FeatureLinks.feature_ids == feature_id).fetch(None)
+  return feature_links if feature_links else []
+
+
+def get_by_feature_id(feature_id: int) -> list[dict[str, Any]]:
+  """Return a list of dicts of FeatureLinks for a given feature id
+  The returned dicts only include the url, type, and information fields.
+  This is used by the api to return json to the client.
+  """
+  feature_links = _get_feature_links(feature_id)
+  return [link.to_dict(include=['url', 'type', 'information']) for link in feature_links]
 
 
 def _index_feature_links(fe: FeatureEntry) -> None:
